generator/requirements.txt
google-cloud-aiplatform
fastapi
uvicorn
gitpython
websockets
